Functions moved to `utils`, new `constants` to support new res easily
